Change semantics of rule references
Previously the reference was [repo]:[name]:[version]. Now the
reference is [namespace]:[name]:[version]. This is how the
documentation for the language was originally written. Somehow [repo]
seemed like a good idea. Now that the namespacing and new relations
b/w Registry and Repo are solidified, it's better to have a namespace.