Removes WHEN clause from REVISE step
The REVISE step functionality tries to do too much. It allows simple
matching with WHENs and it issues Revisions to Documents. Rule authors
already have the tools to perform the manipulations offered by the
WHEN in REVISE. They should use these to bring the table into the
shape they need for issuing the REVISE.

Change-Id: I0a93528cb20ab8984f289d07ca894cb36196c83e