Removes the concept of table references
In all places where a table reference would have been used, a simple name is
used. The interpreter should assume that all tables do not require sections and
simply assume the section is "tables". Therefore, we do not need to all
specifying DIFFERENT sections for tables. This simplifies key references as we
can assume that all sectioned references refer to values in documents.